+++ R/street-crime-outcome.R
@@ -17,16 +17,16 @@
 #' @param date The year and month in "YYYY-MM" form. If `NULL`, latest
 #' available month will be returned.
 #'
-#' @return A tibble with details of street crime outcomes.
+#' @return A `tibble` with details of street crime outcomes.
 #' @export
 #'
 #' @examples
 #' \dontrun{
 #' street_crime_outcome1 <- ukc_street_crime_outcome(location = 883498)
-#' 
+#'
 #' street_crime_outcome2 <- ukc_street_crime_outcome(lat = 52, lng = 0)
 #' }
-#' 
+#'
 ukc_street_crime_outcome <- function(lat, lng, location, date = NULL) {
   date_query <- ukc_date_processing(date)