+++ R/utils-get-data.R
@@ -1,22 +1,45 @@
 
 # data retrieval function
-ukc_get_data <- function(query) {
-  df <- tibble::as_tibble(jsonlite::fromJSON(paste0(baseurl, query)))
+ukc_get_data <- function(query, ...) {
+  x <- httr::GET(paste0(baseurl, query), ...)
 
-  df
+  if (httr::status_code(x) != "200") {
+    message(paste("Request returned error code:", httr::status_code(x)))
+    return(NULL)
+  } else {
+    df <- tibble::as_tibble(
+      jsonlite::fromJSON(httr::content(x, as = "text", encoding = "utf8"),
+        flatten = TRUE
+      )
+    )
+
+    names(df) <- snakecase::to_snake_case(names(df))
+    names(df) <- gsub("location_", "", names(df), fixed = TRUE)
+
+    df
+  }
 }
 
 
 # separate function for specific crime outcomes cause different format in API
-ukc_get_data_specific_crime <- function(query) {
-  api_return <- jsonlite::fromJSON(paste0(baseurl, query))
+ukc_get_data_specific_crime <- function(query, ...) {
+  x <- httr::GET(paste0(baseurl, query), ...)
 
-  if (is.null(api_return$outcomes)) {
-    df <- tibble::as_tibble(purrr::compact(api_return$crime))
+  if (httr::status_code(x) != "200") {
+    message(paste("Request returned error code:", httr::status_code(x)))
+    return(NULL)
   } else {
-    df <- api_return
+    api_return <- jsonlite::fromJSON(httr::content(x, as = "text", encoding = "utf8"),
+      flatten = TRUE
+    )
+
+    if (is.null(api_return$outcomes)) {
+      df <- tibble::as_tibble(purrr::compact(api_return$crime))
+    } else {
+      df <- api_return
+    }
+    df
   }
-  df
 }
 
 # because neighbourhood data is weird