+++ tests/testthat/test-empty-functions.R
@@ -1,7 +1,7 @@
 context("test-update")
 
 test_that("empty-functions works", {
-  skip_on_cran()
+  # skip_on_cran()
   expect_silent(x <- ukc_last_update())
   expect_true(is.character(x$date))
 
@@ -9,4 +9,8 @@ test_that("empty-functions works", {
   expect_length(crimes, 2)
   expect_true(nrow(crimes) == 15)
   expect_true("drugs" %in% crimes$url)
+
+  available <- ukc_available()
+
+  expect_s3_class(available, "data.frame")
 })