+++ tests/testthat/test-specific-crime-no-location-crime.R
@@ -1,13 +1,13 @@
 context("test-specific-crime")
 
-test_that("specific-crime and no-location crime works", {
-  skip_on_cran()
+test_that("specific-crime and no-location crime", {
+  # skip_on_cran()
   no_location <- ukc_crime_no_location(
     force = "city-of-london",
     date = "2019-01"
   )
 
-  expect_length(no_location, 9)
+  expect_true(all(is.na(no_location$location)))
 
   expect_true(
     "a832abdef7dc2a9a794ca3ce9730e541619c316e6cc87132334a50ec7762b8ae" %in%
@@ -16,7 +16,7 @@ test_that("specific-crime and no-location crime works", {
 
   outcome <- ukc_specific_outcome(
     "a832abdef7dc2a9a794ca3ce9730e541619c316e6cc87132334a50ec7762b8ae"
-    )
+  )
 
   expect_length(outcome, 2)
 
@@ -27,6 +27,12 @@ test_that("specific-crime and no-location crime works", {
   expect_length(outcome2, 2)
   expect_true(is.list(outcome2))
 
+
+  expect_message(
+    ukc_specific_outcome("asdffsda"),
+    "Request returned error code: 404"
+  )
+
   no_location2 <- ukc_crime_no_location(
     force = "city-of-london",
     date = "2019-01",
@@ -34,4 +40,7 @@ test_that("specific-crime and no-location crime works", {
   )
 
   expect_true(all(no_location2$category == "drugs"))
+
+  expect_error(ukc_crime_no_location())
+  expect_error(ukc_specific_outcome())
 })